Adds a command handeling library
This library is meant to intercept the serial communication and handel
commands sent over it. Due to this interception is means the serial
library should not be used for any other purpose as it will interfere
with the communications between the host and the robot. All
communications should be done through this library.